+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02. functions and logic/core/core.get_party_type_id_by_party_code.sql
@@ -1,5 +1,5 @@
 CREATE FUNCTION core.get_party_type_id_by_party_code(text)
-RETURNS smallint
+RETURNS integer
 AS
 $$
 BEGIN
