+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02. functions and logic/core/core.get_unit_id_by_unit_code.sql
@@ -1,5 +1,5 @@
 CREATE FUNCTION core.get_unit_id_by_unit_code(text)
-RETURNS smallint
+RETURNS integer
 AS
 $$
 BEGIN
